requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
amqp==5.1.1
async-generator==1.10
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.2.0
backports.entry-points-selectable==1.1.1
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.11.1
billiard==3.6.4.0
blinker==1.4
Brotli==1.0.9
bs4==0.0.1
CacheControl==0.12.10
cachy==0.3.0
celery==5.2.7
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.8
cleo==0.8.1
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clikit==0.6.2
colorama==0.4.4
colored==1.4.4
configparser==5.1.0
crashtest==0.3.1
crayons==0.4.0
cryptography==36.0.0
Deprecated==1.2.13
distlib==0.3.6
et-xmlfile==1.1.0
execnet==1.9.0
Faker==9.8.3
filelock==3.8.0
future==0.18.2
h11==0.12.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
hyperframe==6.0.1
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
install==1.3.5
itsdangerous==2.1.2
jeepney==0.8.0
Jinja2==3.1.1
jsonschema==4.17.3
kaitaistruct==0.9
keyring==21.8.0
kombu==5.2.4
lockfile==0.12.2
MarkupSafe==2.1.1
msgpack==1.0.3
numpy==1.23.5
openpyxl==3.0.9
outcome==1.1.0
packaging==20.9
pandas==1.5.2
pastel==0.2.1
pexpect==4.8.0
pkginfo==1.8.1
platformdirs==2.4.0
pluggy==1.4.0
poetry==1.1.11
poetry-core==1.0.7
prompt-toolkit==3.0.30
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
pydivert==2.1.0
pylev==1.4.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyrsistent==0.20.0
PySocks==1.7.1
pytest==8.0.0
pytest-check==2.3.1
pytest-forked==1.4.0
pytest-rerunfailures==10.2
pytest-sugar==0.9.4
pytest-xdist==2.5.0
python-dateutil==2.8.2
python-decouple==3.5
python-dotenv==0.20.0
pytz==2022.2.1
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.26.0
requests-toolbelt==0.9.1
SecretStorage==3.3.3
selenium==4.1.3
shellingham==1.4.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
tomlkit==0.7.2
tqdm==4.64.1
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.0.1
urllib3==1.26.7
urllib3-secure-extra==0.1.0
vine==5.0.0
virtualenv==20.10.0
wcwidth==0.2.5
webdriver-manager==4.0.0
webencodings==0.5.1
Werkzeug==2.1.1
wrapt==1.14.1
wsproto==1.0.0
zipp==3.8.0
